### 2019-03-21 #47 @whedon
[PRE REVIEW]: FDS Python: Fundamentals and Data Structures with Python
**Submitting author:** @rambasnet (<a href="http://orcid.org/0000-0001-6864-6893">Ram Basnet</a>)
**Repository:** <a href="https://github.com/rambasnet/thinkpythonnotebooks" target ="_blank">https://github.com/rambasnet/thinkpythonnotebooks</a>
**Version:** v1.0.0
