### 2019-03-21 #47 @labarba
👋  @rambasnet — Thanks for your submission! Could you move your bibliography to a `.bib` file? The auto-compilation expects this. Check the [Author Guidelines](http://jose.theoj.org/about#author_guidelines)
