### 2019-04-16 #47 @rambasnet
@labarba -- Just moved references to .bib file. All should be good to go.
