### 2019-04-27 #47 @labarba
@mcburton 👋 — Would you be willing to contribute a review for this JOSE submission? (We already have one reviewer.) "FDS Python: Fundamentals and Data Structures with Python" https://github.com/rambasnet/thinkpythonnotebooks
