### 2019-05-13 #52 @labarba
👋  @AllenDowney, @mcburton — We've started the review! This is where it happens. You each have a review checklist at the top of this issue, to guide you through. The version item is something we can check off in the end, as we will ask the authors to do a release after any revisions. Statement of need should be a section in the paper (the checklist has it also in the Documentation, but that's optional).

Feel free to ask any questions here, and @katyhuff will help you out. You can also post questions to the author, or open issues in the submission repository.

Please tick off your Conflict of Interest and Code of Conduct items ASAP, so we all see the review has started. Thanks!
