### 2019-06-04 #52 @AllenDowney
Under General Checks:

* Version: Does the release version given match the repository release (v1.0.0)?

I don't see a version number in the repo.

* Authorship: Has the submitting author (@rambasnet) made visible contributions to the module?

Yes.

* Does the full list of authors seem appropriate and complete?

It looks like there is just one author, and only one contributor to the repo, so I guess I'll check this box.
