### 2019-06-04 #52 @AllenDowney
It looks like these notebooks are based on *How to Think Like a Computer Scientist: Learning with Python 3 (RLE)*, but the documentation refers to *Think Python*.  These are different books (although they share a common ancestor).  Ideally, the paper and documentation should refer to *How to Think*, and remove references to *Think Python*.
