### 2019-06-04 #52 @katyhuff
> I don't see a version number in the repo.

To create the DOI on submission, the author will need to create a release. Typically, this is done before the JOSS submission. However, if the authors prefer, sometimes the review is merely conducted on the most recent commit to the default branch and the version (in this case (1.0.0)) is assigned at the review completion when the software DOI is assigned.

> It looks like there is just one author, and only one contributor to the repo, so I guess I'll check this box.

Yes, unless there's reason to believe other authors exist, we do go by commits.

> It looks like these notebooks are based on How to Think Like a Computer Scientist: Learning with Python 3 (RLE), but the documentation refers to Think Python. These are different books (although they share a common ancestor). Ideally, the paper and documentation should refer to How to Think, and remove references to Think Python.

@rambasnet please note the above comment.
