### 2019-07-10 #52 @rambasnet
Hi, @labarba @katyhuff @AllenDowney,
 I was just curious about the review status as I've not noticed any progress since the last comment. Are reviewers waiting for my response on me for anything that I missed? Thanks in advance!
