### 2019-07-11 #52 @katyhuff
Thank you for pinging this item @rambasnet .

@mcburton & @AllenDowney : Can you estimate a time of completion for your reviews? I recognize that summer (e.g. the SciPy conference this week) may impede progress somewhat. Please let us know what timeline will be reasonable for your review completion.
