### 2019-07-26 #52 @labarba
@AllenDowney, @mcburton — hi folks! May we have an update from you on the status of your review? Thank you! 🙏
