### 2019-07-26 #52 @AllenDowney
@rambasnet The installation instructions look good, but it doesn't look like there is a list of requirements.  Consider adding requirements.txt or environment.yml (although it looks like the notebooks run on Binder, so maybe you have no requirements other than Python and Jupyter).
