### 2019-07-26 #52 @rambasnet
@AllenDowney Yes, all it needs is Python 3 and Jupyter Notebook.
