### 2019-07-26 #52 @AllenDowney
@rambasnet On the question of Content quality, I think the primary contribution here is porting the code examples and exercises from the book into notebooks.  Currently, the text in the notebooks is minimal, so they would not work well on their own, but I understand they are mean to supplement the book.  Over time, it might be nice to flesh out the text in the notebooks.
