### 2019-07-26 #52 @AllenDowney
@rambasnet It looks like the references in paper.md are missing.  And the format is a little strange, but I don't know if there's a template it's supposed to follow.
