### 2019-07-26 #52 @rambasnet
@AllenDowney, I agree with your comment on the content bing minimal which is intentional as the goal is not to compliment the text but to supplement. Also, research shows students do not always read text and we hope that these notebooks will force them to read.
References are in paper.bib file that is how the editorial wants.
