### 2019-07-26 #52 @labarba
Reading this thread, I'm concerned that this project may not be quite what JOSE is looking for. We seek to publish computational learning modules that are ready for adoption by other researchers, and are complete in that sense. Maybe a case can be made, since the textbook these notebooks complement is openly available. But spartan notebooks made up of mostly code in general would not be sufficient for a JOSE publication.
