### 2019-07-26 #52 @rambasnet
main contributions:
- all the important keywords, terms, topics, and definitions used in typical CS1 and CS2 courses are included in easy to read fashion making it easier for instructors to use during class (instead of jumping between powerpoint/text and code editor). This, of course, is mostly based on the original text with minor addition and or tweak where applicable. Every chapter has a direct link to corresponding online chapter in the textbook and other resources e.g. from python documentation. We don't see a value in completely replicating a great textbook.
- code visualization using http://pythontutor.com/ where appropriate right in Notebook or on the site with pre filled code thus saving time for instructors to work on many examples (text doesn't provide this feature to the best of my knowledge)
- include challenging and fun problems where applicable from https://open.kattis.com -- repo of programming contest problems (text doesn't provide this feature)

I hope this clarifies some of the doubt you may have about the importance and contributions of these notebooks in teaching problem solving using Python using hands-on approach.
