### 2019-07-28 #52 @labarba
@rambasnet — I've had a browse into your notebooks, and read the paper, and as the handling editor, I do maintain several concerns about the submission. It doesn't seem ready to be published. Generally, we'd expect fully narrated and ready-to-use lessons. I do recognize that the purpose and design of this collection of notebooks is not a "learning module" using computing to learn by the JOSE expectation, due to the fact that this set supplements an existing open textbook.

However, even if we consider the notebooks as _dependent on_ the textbook, they still look unfinished. For example, the first notebook has a link to the corresponding chapter in the book, but the next few (I didn't click through all of them) do not. The reader needs to use the README as a guide.

In the paper, the Statement of Need is supposed to address the need for the content submitted here. You do not need to discuss the need for Python or Jupyter notebooks, but the need for _the materials you created_.

Given that this submission is unlike others in JOSE (not being self-contained), it is more important to take the time to explain to readers why they might need this content and how they might use it.

Please give some thought to these comments, and let us know if you think you can work on your submission to make it closer to a "finished product."
