### 2019-07-29 #52 @rambasnet
@labarba - Thank you for the feedback!
We'll continue to use these notebooks in our classrooms and improve them with your suggestions in mind to make it closer to a "finished product."
