### 2019-07-30 #52 @labarba
@AllenDowney, @mcburton — I'm going to pause the review, in light of the exchange with the author just above ☝️
This collection of notebooks needs a bit of work before they are in publishable form. We are very grateful for the time you've contributed to this review so far, and hope you'll be willing to come back to it when the authors revive their submission 🙏
