### 2019-07-30 #52 @AllenDowney
Ok, thanks, Lorena.
