### 2019-11-10 #52 @rambasnet
Hi Lorena, @labarba I've cleaned up and reorganized the notebooks quite a bit while using them in two courses beginning and advanced programming with Python this semester. I was hoping if I could get the review resumed on this submission.
Thank you for your time!
