### 2019-11-10 #52 @labarba
👋 @AllenDowney, @mcburton — The author of this submission has checked in and reports a refactoring of the materials. Are you willing to step back into reviewing this for JOSE? I see that you both had made lots of progress already.
