### 2020-02-03 #52 @mcburton
I am excited to see the direction that these materials are going, but I think there is still more work to be done clarifying how these are to be used and how much explanatory content should be included in the notebooks.

## Major Comments
* The `paper.md` indicates these notebooks are being used at Colorado Mesa University, but additional details specifically *how* they are being used would strengthen the paper. How are these notebooks intended to be used with the book? The notebook chapters don’t fully correspond to chapters in the book, especially for new content.
* There needs to be some work distinguishing these materials from the ThinkCSPy book. Much of the content in these notebooks is derived directly from the book, but then additional chapters, like the chapter on sqliteDB appears to be fundamentally new.
* The new content is very sparse and includes very little discussion both for students and teachers besides the code and headers. With few pointers for how it aught to be used I’m not sure these notebooks are ready for someone to just pick up and use. This may be less of a problem for the notebooks at accompany the book, but is challenging for the new content.
* The content in each notebook would be strengthened with an explicit set of learning objectives at the beginning and some summative statements of what was learned at the end of each notebook.

## Minor changes
* The `paper.md` is still missing references. They are in the bib file but should be included in a human readable format in  `paper.md`
* There is a minor markdown syntax error in the statement of need header in `paper.md`
*  The table of contents ordering is not in sync with notebook ordering in the repository starting with chapter 20.

In summary, I think @rambasnet has done some work to address the concerns that @AllenDowney and @labarba had raised in earlier comments in this thread, but I don't think the changes are complete. These materials need additional work in terms of expository comments within the individual notebooks. I understand @rambasnet's desire to not write a textbook, but if the idea of these notebooks is to be sparse then there needs to be additional content for the instructor (instructor's notes perhaps?) that provide some guidance about how these materials are to be used. The ThinkCSPy book does this for some of the chapters, but these materials don't have a 1 to 1 relationship with that book.

I think there is a sweet spot between sparse notebooks filled with code and a full textbook. Some additional specific details about *how* these materials are intended to be used may help clarify and set expectations.

If you have any questions please let me know.
