### 2020-03-26 #52 @rambasnet
Thank you for the review @mcburton. I appreciate your feedback.

I've a strong feeling that my intentions of these notebooks will not meet the requirements of the JOSE journal. So, I, respectfully, would like to withdraw my submission.

Thank you all @labarba and @AllenDowney for your valuable time to give this submission a chance.
