### 2020-03-27 #52 @labarba
No problem @rambasnet. Thank you for your efforts and considering JOSE to showcase your work.

I will now withdraw your submission.

Many thanks to @AllenDowney and @mcburton for the time invested in reviewing this work 🙏
